% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mids.R
\docType{data}
\name{mids}
\alias{mids}
\title{Corrected mass isotopomer distributions}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 44499 rows and 10 columns.
}
\usage{
mids
}
\description{
A data set containing corrected mass isotope distributions. Technical
replicates have been filtered from the data.
}
\details{
\describe{
\item{method}{
\code{fs} = full scan \cr
\code{sim} = selected ion monitoring}
\item{cell_type}{
\code{lf} = lung fibroblasts \cr
\code{pasmc} = pulmonary artery smooth muscle cells}
\item{date}{start date of an experiment}
\item{tracer}{
\code{glc2} = [1,2-13C2] glucose \cr
\code{glc6} = [U-13C6] glucose \cr
\code{q5} = [U-13C5] glutamine \cr
\code{lac3} = [U-13C3] lactate}
\item{oxygen}{ambient oxygen level}
\item{treatment}{
\code{none} = no treatment \cr
\code{DMSO} = 0.1\% DMSO \cr
\code{BAY} = 10 μM molidustat}
\item{time}{hours}
\item{metabolite}{name of measurement}
\item{isotope}{mass shift of detected isotope}
\item{mid}{mole fraction of the indicated isotope}
}
}
\keyword{datasets}
